Split components and containers into separate folders.
