+++ assets/js/careeropportunity/components/Job.js
@@ -1,0 +1,15 @@
+class Job extends React.Component {
+  constructor() {
+    super();
+  }
+
+  render() {
+    return (
+      <div className="job">
+        {this.props.jobTitle}
+      </div>
+    );
+  }
+}
+
+export default Job;
