+++ assets/js/careeropportunity/index.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import ReactDom from 'react-dom';
+import JobTypeButtonsContainer from './containers/JobTypeButtonsContainer';
+import JobsContainer from './containers/JobsContainer.js';
 
 let data = {
   jobs: [{
@@ -25,144 +27,6 @@ let data = {
   }]
 };
 
-class Job extends React.Component {
-  constructor() {
-    super();
-  }
-
-  render() {
-    return (
-      <div className="job">
-        {this.props.jobTitle}
-      </div>
-    );
-  }
-}
-
-class JobList extends React.Component {
-  constructor(props) {
-    super();
-
-    this.state = {
-      selectedJobTypes: {}
-    };
-
-    this.handleJobTypeChange = this.handleJobTypeChange.bind(this);
-  }
-
-  handleJobTypeChange(selectedJobTypes) {
-    this.setState({
-      selectedJobTypes: selectedJobTypes
-    });
-  };
-
-  render() {
-    // Warning: Will be true on the initial render, which is the reason as to why we can
-    // have an empty object in the initial selectedJobTypes state!
-    let allJobTypesDisabled = true;
-
-    for (let key in this.state.selectedJobTypes) {
-      if (this.state.selectedJobTypes.hasOwnProperty(key) && this.state.selectedJobTypes[key] === true) {
-        allJobTypesDisabled = false;
-      }
-    }
-
-    let jobs = this.props.jobs.map(function(job) {
-      if (this.state.selectedJobTypes[job.type] || allJobTypesDisabled) {
-        return (
-          <Job jobTitle={job.title} />
-        );
-      }
-    }.bind(this));
-
-    return (
-      <div class="job-list">
-        {jobs}
-      </div>
-    );
-  }
-}
-
-class JobType extends React.Component {
-  constructor() {
-    super();
-
-    this.handleClick = this.handleClick.bind(this);
-  }
-
-  handleClick(key) {
-    this.props.handleChange(key);
-  }
-
-  render() {
-    return (
-      <button onClick={this.handleClick.bind(this, this.props.jobType)}>{this.props.title} - {this.props.selected ? 'yes' : 'no'}</button>
-    );
-  }
-}
-
-class JobTypeMenu extends React.Component {
-  constructor(props) {
-    super();
-
-    this.state = {
-      selectedJobTypes: {}
-    };
-
-    for (let key in props.jobTypes) {
-      if (props.jobTypes.hasOwnProperty(key)) {
-        this.state.selectedJobTypes[key] = false;
-      }
-    }
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleReset = this.handleReset.bind(this);
-  }
-
-  handleChange(selectedKey) {
-    let selectedJobTypes = this.state.selectedJobTypes;
-
-    selectedJobTypes[selectedKey] = !selectedJobTypes[selectedKey];
-
-    this.setState({
-      selectedJobTypes: selectedJobTypes
-    });
-
-    this.props.handleJobTypeChange(selectedJobTypes);
-  }
-
-  handleReset() {
-    let resetJobTypes = {};
-
-    for (let key in this.state.selectedJobTypes) {
-      if (this.state.selectedJobTypes.hasOwnProperty(key)) {
-        resetJobTypes[key] = false;
-      }
-    }
-
-    this.setState({
-      selectedJobTypes: resetJobTypes
-    });
-
-    this.props.handleJobTypeChange(resetJobTypes);
-  }
-
-  render() {
-    let jobTypeButtons = this.props.jobTypes.map(function(jobType) {
-      return (
-        <JobType selected={this.state.selectedJobTypes[jobType.key]} jobType={jobType.key} handleChange={this.handleChange} title={jobType.title} />
-      );
-    }.bind(this));
-
-    return (
-      <div>
-        {jobTypeButtons}
-        <button onClick={this.handleReset}>reset</button>
-      </div>
-    );
-  }
-}
-
 class App extends React.Component {
   constructor(data) {
     super();
@@ -186,11 +50,11 @@ class App extends React.Component {
           </div>
 
           <div className="col-md-8">
-            <JobList jobs={this.props.data.jobs} handleJobTypeChange={this.handleJobTypeChange} ref="jobList" />
+            <JobsContainer jobs={this.props.data.jobs} handleJobTypeChange={this.handleJobTypeChange} ref="jobList" />
           </div>
 
           <div className="col-md-4">
-            <JobTypeMenu jobTypes={this.props.data.jobTypes} handleJobTypeChange={this.handleJobTypeChange} />
+            <JobTypeButtonsContainer jobTypes={this.props.data.jobTypes} handleJobTypeChange={this.handleJobTypeChange} />
           </div>
       </div>
     );
